requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-telegram-bot==21.3
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
pytest-httpx==0.30.0
